+++ opentelemetry-jaeger/Cargo.toml
@@ -34,6 +34,10 @@ tokio = { version = "0.2", features = ["udp", "sync"], optional = true }
 wasm-bindgen = { version = "0.2", optional = true }
 wasm-bindgen-futures = { version = "0.4.18", optional = true }
 thiserror = "1.0"
+lazy_static = "1.4"
+
+[dev-dependencies]
+opentelemetry = { version = "0.10", default-features = false, features = ["trace", "testing"], path = "../opentelemetry" }
 
 [dependencies.web-sys]
 version = "0.3.4"
