+++ opentelemetry-jaeger/src/exporter/transport/buffer.rs
