+++ opentelemetry-jaeger/src/exporter/transport/noop.rs
