+++ opentelemetry-zipkin/src/exporter/uploader.rs
@@ -1,6 +1,6 @@
 //! # Zipkin Span Exporter
-use crate::model::span::Span;
-use crate::Error;
+use crate::exporter::model::span::Span;
+use crate::exporter::Error;
 use http::{header::CONTENT_TYPE, Method, Request, Uri};
 use opentelemetry::sdk::export::trace::{ExportResult, HttpClient};
 use std::fmt::Debug;
